api/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pystray>=0.19.5
Pillow>=10.3.0
python-multipart>=0.0.9
httpx>=0.27.0
pytest>=8.1.0
flake8>=7.0.0
